requirements.txt
# Core data
pandas==2.2.2
numpy==1.26.4
joblib

# NLP & ML
scikit-learn==1.4.2
sentence-transformers==2.7.0

# PyTorch — CUDA 12.1 wheel (RTX 4060 / any CUDA-capable GPU)
# Install via: pip install torch==2.2.2 --index-url https://download.pytorch.org/whl/cu121
# For CPU-only fallback: pip install torch==2.2.2 --index-url https://download.pytorch.org/whl/cpu
torch==2.2.2

# Nomic-embed-text-v1.5 requires einops for its custom model code
einops

# Web scraping
requests==2.31.0
beautifulsoup4==4.12.3
playwright==1.58.0            # Jadarat.sa collector (requires: python -m playwright install chromium)
python-jobspy==1.1.82         # LinkedIn / Indeed / Glassdoor real-job collector

# PDF parsing
pdfplumber==0.11.0

# UI
streamlit==1.35.0

# REST API backend (for Next.js frontend)
fastapi==0.135.3
uvicorn==0.43.0
python-multipart==0.0.9

# Testing
pytest==8.2.2
